Move tooltips and popovers
