+++ shiny/ui/_utils.py
@@ -12,6 +12,8 @@ from htmltools import (
     tags,
 )
 
+from .. import Session
+from ..session import require_active_session
 from ..types import MISSING, MISSING_TYPE
 
 
@@ -66,3 +68,10 @@ def _find_child_strings(x: TagList | TagNode) -> str:
     if isinstance(x, str):
         return x
     return ""
+
+
+def _session_on_flush_send_msg(
+    id: str, session: Session | None, msg: dict[str, object]
+) -> None:
+    session = require_active_session(session)
+    session.on_flush(lambda: session.send_input_message(id, msg), once=True)
